% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LREP-package.R
\docType{package}
\name{LREP-package}
\alias{LREP}
\alias{LREP-package}
\title{LREP: Estimate and Test Exponential vs. Pareto Distributions}
\description{
The programs were developed for estimation of parameters and testing exponential versus Pareto distribution during our work on hydrologic extremes. See Kozubowski, T.J., A.K. Panorska, F. Qeadan, and A. Gershunov (2007) <doi:10.1080/03610910802439121>, and Panorska, A.K., A. Gershunov, and T.J. Kozubowski (2007) <doi:10.1007/978-0-387-34918-3_26>.
}
\details{
We consider the problem of maximum likelihood estimation of the parameters of the Pareto Type II (Lomax) distribution. We show that in certain parametrization and after modification of the parameter space to include exponential distribution as a special case, the MLEs of parameters always exist. Moreover, the MLEs have a non standard asymptotic distribution in the exponential case due to the lack of regularity. Further, we develop a likelihood ratio test for exponentiality versus Pareto II distribution.
}
\references{
\enumerate{
\item Kozubowski, T.J., A.K. Panorska, F. Qeadan, and A. Gershunov (2007). Testing exponentiality versus Pareto distribution via likelihood ratio, preprint in review.
\item Panorska, A.K., A. Gershunov, and T.J. Kozubowski (2007). From diversity to volatility: Probability of daily precipitation and extremes. Nonlinear Dynamics in Geosciences (A. Tsonis and J. Elsner, Eds.), Springer, New York, pp 465-484.
}
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/jiqiaingwu/LREP/issues}
}

}
\author{
\strong{Maintainer}: Jiqiang Wu \email{charles.wu@utah.edu}

Authors:
\itemize{
  \item Fares Qeadan \email{fares.qeadan@utah.edu}
  \item Tomasz Kozubowski \email{tkozubow@unr.edu}
  \item Anna Panorska \email{ania@unr.edu}
}

}
