% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmaalphaLREP.R
\name{sigmaalphaLREP}
\alias{sigmaalphaLREP}
\title{The program will take your data as input and return three numbers as output: estimate of s, estimate of alpha, and value of the statistic L.}
\usage{
sigmaalphaLREP(x, tolerance)
}
\arguments{
\item{x}{Import Raw data}

\item{tolerance}{Tolerance quantity}
}
\value{
Output three estimates by using maximum likelihood method
}
\description{
The program will take your data as input and return three numbers as output: estimate of s, estimate of alpha, and value of the statistic L.
}
\section{Fields}{

\describe{
\item{\code{s.hat}}{Output estimate of s}

\item{\code{a.hat}}{estimate of alpha}

\item{\code{log.like.ratio}}{value of the statistic L}
}}

\examples{
x<-rexp(1000,0.1)
print(sigmaalphaLREP(x,10^-12))
}
