% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LREP-package.R
\docType{package}
\name{LREP-package}
\alias{LREP}
\alias{LREP-package}
\title{LREP: Estimate and Test Exponential vs. Pareto Distributions}
\description{
The programs were developed for estimation of parameters and testing exponential versus Pareto distribution during our work on hydrologic extremes.
}
\references{
\enumerate{
\item Kozubowski, T.J., A.K. Panorska, F. Qeadan, and A. Gershunov (2007). Testing exponentiality versus Pareto distribution via likelihood ratio, preprint in review.
\item Panorska, A.K., A. Gershunov, and T.J. Kozubowski (2007). From diversity to volatility: Probability of daily precipitation and extremes. Nonlinear Dynamics in Geosciences (A. Tsonis and J. Elsner, Eds.), Springer, New York, pp 465-484.
}
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/jiqiaingwu/LREP/issues}
}

}
\author{
\strong{Maintainer}: Jiqiang Wu \email{charles.wu@utah.edu}

Authors:
\itemize{
  \item Fares Qeadan \email{fares.qeadan@utah.edu}
  \item Tomasz Kozubowski \email{tkozubow@unr.edu}
  \item Anna Panorska \email{ania@unr.edu}
}

}
